Using react hooks
