+++ web/src/app/alerts/components/AlertDetails.js
@@ -2,15 +2,10 @@ import React, { Component } from 'react'
 import p from 'prop-types'
 import Card from '@material-ui/core/Card'
 import CardContent from '@material-ui/core/CardContent'
-import Divider from '@material-ui/core/Divider'
 import FormControlLabel from '@material-ui/core/FormControlLabel'
 import Grid from '@material-ui/core/Grid'
 import Hidden from '@material-ui/core/Hidden'
-import List from '@material-ui/core/List'
-import ListItem from '@material-ui/core/ListItem'
-import ListItemText from '@material-ui/core/ListItemText'
 import Switch from '@material-ui/core/Switch'
-import Button from '@material-ui/core/Button'
 import Table from '@material-ui/core/Table'
 import TableBody from '@material-ui/core/TableBody'
 import TableCell from '@material-ui/core/TableCell'
@@ -19,7 +14,6 @@ import TableRow from '@material-ui/core/TableRow'
 import Typography from '@material-ui/core/Typography'
 import withStyles from '@material-ui/core/styles/withStyles'
 import isFullScreen from '@material-ui/core/withMobileDialog'
-import moment from 'moment'
 import Countdown from 'react-countdown-now'
 import { Link } from 'react-router-dom'
 import { ScheduleLink, ServiceLink, UserLink } from '../../links'
@@ -28,154 +22,10 @@ import Options from '../../util/Options'
 import gql from 'graphql-tag'
 import PageActions from '../../util/PageActions'
 import Markdown from '../../util/Markdown'
-import { graphql } from 'react-apollo'
+import AlertDetailLogs from './AlertDetailLogs'
 
 const localStorage = window.localStorage
 const exactTimesKey = 'show_exact_times'
-
-const LIMIT = 149
-
-const query = gql`
-  query getAlert($id: Int!, $input: AlertRecentEventsOptions) {
-    alert(id: $id) {
-      data: recentEvents(input: $input) {
-        nodes {
-          timestamp
-          message
-        }
-        pageInfo {
-          hasNextPage
-          endCursor
-        }
-      }
-    }
-  }
-`
-
-@graphql(query, {
-  options: props => {
-    return {
-      showExact: props.showExact,
-      variables: {
-        id: props.data,
-        input: {
-          after: '',
-          limit: 35,
-        },
-      },
-    }
-  },
-  props: props => {
-    return {
-      showExact: props.ownProps.showExact,
-      data: props.data,
-      loadMore: queryVariables => {
-        return props.data.fetchMore({
-          variables: queryVariables,
-          updateQuery(prev, { fetchMoreResult }) {
-            if (!fetchMoreResult) return prev
-            return {
-              alert: {
-                ...fetchMoreResult.alert,
-                data: {
-                  ...fetchMoreResult.alert.data,
-                  nodes: prev.alert.data.nodes.concat(
-                    fetchMoreResult.alert.data.nodes,
-                  ),
-                },
-              },
-            }
-          },
-        })
-      },
-    }
-  },
-})
-@withStyles(styles)
-@isFullScreen()
-class AlertLogs extends Component {
-  handleLoadMore = () => {
-    let id = this.props.data.variables.id
-    let after = this.props.data.alert.data.pageInfo.endCursor
-    this.props.loadMore({
-      id: id,
-      input: {
-        after: after,
-        limit: LIMIT,
-      },
-    })
-  }
-  renderNoResults = () => {
-    return (
-      <ListItem>
-        <ListItemText primary='No events.' />
-      </ListItem>
-    )
-  }
-  renderLoadMore(hasNextPage) {
-    if (hasNextPage) {
-      return (
-        <Button
-          style={{ width: '100%' }}
-          onClick={this.handleLoadMore}
-          variant='outlined'
-          data-cy='load-more-logs'
-        >
-          Load More
-        </Button>
-      )
-    }
-  }
-  render() {
-    const { alert } = this.props.data
-    let data = []
-    if (alert) {
-      if (alert.data.nodes.length <= 35) {
-        this.props.data.startPolling(2500)
-      } else this.props.data.stopPolling()
-      data = alert.data ? alert.data.nodes : []
-    }
-
-    if (!data.length) {
-      return this.renderNoResults()
-    }
-    if (this.props.showExact) {
-      data = data.map((log, index) => (
-        <div key={index}>
-          <Divider />
-          <ListItem>
-            <ListItemText
-              primary={moment(log.timestamp)
-                .local()
-                .format('MMM Do YYYY, h:mm:ss a')}
-              secondary={log.message}
-            />
-          </ListItem>
-        </div>
-      ))
-    } else {
-      data = data.map((log, index) => (
-        <div key={index}>
-          <Divider />
-          <ListItem>
-            <ListItemText
-              primary={moment(log.timestamp)
-                .local()
-                .calendar()}
-              secondary={log.message}
-            />
-          </ListItem>
-        </div>
-      ))
-    }
-    return (
-      <List style={{ padding: 0 }}>
-        {data}
-        {this.renderLoadMore(alert.data.pageInfo.hasNextPage)}
-      </List>
-    )
-  }
-}
 
 @withStyles(styles)
 @isFullScreen()
@@ -238,9 +88,9 @@ export default class AlertDetails extends Component {
           className={this.props.classes.tableCardContent}
           style={{ paddingBottom: 0 }}
         >
-          <AlertLogs
-            data={this.props.data.number}
-            showExact={this.state.showExactTimes}
+          <AlertDetailLogs
+            alertID={this.props.data.number}
+            showExactTimes={this.state.showExactTimes}
           />
         </CardContent>
       </Card>
